Update and move project verification to ConfigDB
This removes the project verification from the project generator and moves
it to the ConfigDB as a sanity check.
The verification earlier checked for the presence of 1 application module
and 1 board. The update makes it check for _no more than 1_ of each, so
that it allows for 0 application modules or boards.