Make Studio proj.gen. handle missing URL for proj.
This prevents the Atmel Studio project generator from causing an exception
in cases where a documentation URL for the project cannot be found, i.e.,
None is returned from get_help_url().

Previously, this would be passed on to the parser, which would then raise
an exception when trying to write the XML file.