Make url + path attr. names consistent in ext.XML
This makes the url and path attribute names for documentation linking
consistent in the extension XML schema, updating the current extension.xml
files for ASF and unit tests.
The change is that the index-page elements now have "path" and "url" for
offline and online documentation, respectively, while the module-level
elements have "basepath" and "baseurl" since they specify the base on
which the specific location shall be built.