Make linker script etc. getter extension-friendly
This makes the generic function for fetching build items from either the
current project or an optional selector more extension-friendly, by
allowing the optional selector to be defined in either the project's home
database or in ASF.

If a selector is specified and no build item of specified type found in
the project, the function will try to find the selector in the current
database. If that fails, it will try to find a dependency on ASF in the
current extension, and if it exists, try to find the selector in the
corresponding ASF database.