Fix up display of macro definitions in generated PDF documentation.
Add a new "Macro" section to the PDF output, and populate it with
formatted versions of macro definitions in the Doxygen documentation. This
alters the current XPATH queries for the function documentation to only
include custom sections with function definitions, and adds a new macro
section for custom sections with macro definitions.

Macros are formatted as a table, one table per macro showing the
initializer and description.

NB: A different custom section should be used for macros and for functions
    - if both types appear in a single custom section, the complete section
    contents will be repeated in both portions of the PDF.