Ensure the master group reference is created in the documentation.
Map the main group's reference ID to the start of the documentation, so
that references back to the group work in the generated PDF
documentations. Without this patch, references back to the main group of a
driver/application/etc. group will fail in the PDF documentation.

Add description for custom sections to the PDF output, so that
descriptions for a custom grouping will be user visible.