Fix up PDF generation of type definitions and public variables.
Fix up the XML transformer to correct the way that type definitions and
variables appear in the PDF output. Remove incorrect variable display
templates and replace them with a new template that can process both
variables and type definitions. A bug in the current Doxygen causes
function pointer type definitions to be mapped as variables, thus the two
need to be grouped together in the documentation to prevent glitches.

Fix up the way macros are displayed in the documentation - use a C snippet
to show the documentation rather than a table.