Remove toolchain config content from .cproj templates
This removes content of the toolchain config tags in the .cproj template
files, to ensure that all toolchain config content comes from the ASF XML
and project generator.

Technically, the tags can be removed altogether, but they are left in the
templates to make it easier to compare the generated project files before
and after the fix for FDK-464.