Fix SAM CMSIS support for the SAM3SD devices.
Fix up the SAM CMSIS module so that the generic SAM3SD devices are
supported - without this, the SAM3SD User Board template would not work
correctly. Note that the SAM3S8 has no official support in ASF as it is
part of the same CMSIS module ass the SAM3SD8, but has no device support;
adding device support for this will cause multiple CMSIS modules to be
compatible for this device and will wreak havok. Need to discuss with ASF
team how to approach this.